+++ twisted/plugins/txtorcon_endpoint_parser.py
@@ -1,2 +1,3 @@
 import txtorcon
 tcpHiddenServiceEndpointParser = txtorcon.TCPHiddenServiceEndpointParser()
+tcpTorClientEndpointParser = txtorcon.TorClientEndpointStringParser()